refactor: move headers to their own distinct directory
