+++ js-bindings/wrappers/BignumWrapper.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "../helpers.h"
+#include <helpers.h>
 #include "BignumWrapper.h"
 
 namespace js_wrappers {
