+++ python-bindings/CMakeLists.txt
@@ -7,5 +7,11 @@ FetchContent_Declare(
 )
 FetchContent_MakeAvailable(pybind11)
 
+include_directories(
+        ${INCLUDE_DIRECTORIES}
+        ${CMAKE_CURRENT_SOURCE_DIR}
+        ${CMAKE_CURRENT_SOURCE_DIR}/../include
+)
+
 pybind11_add_module(blspy ${CMAKE_CURRENT_SOURCE_DIR}/pythonbindings.cpp)
 target_link_libraries(blspy PRIVATE dashbls)
