+++ src/CMakeLists.txt
@@ -1,4 +1,4 @@
-file(GLOB HEADERS ${CMAKE_CURRENT_SOURCE_DIR}/*.hpp)
+file(GLOB HEADERS ${CMAKE_CURRENT_SOURCE_DIR}/../include/dashbls/*.hpp)
 source_group("SrcHeaders" FILES ${HEADERS})
 
 add_library(dashbls
@@ -18,6 +18,7 @@ target_include_directories(dashbls
   PUBLIC
     ${CMAKE_CURRENT_SOURCE_DIR}
     $<$<BOOL:${GMP_FOUND}>:${GMP_INCLUDES}>
+    ${CMAKE_CURRENT_SOURCE_DIR}/../include/dashbls
     ${CMAKE_CURRENT_SOURCE_DIR}/../depends/relic/include
     ${CMAKE_CURRENT_BINARY_DIR}/../depends/relic/include
 )
